# public/swagger.yaml
openapi: 3.0.0
info:
  title: Flight Booking API
  version: 1.0.0
  description: API documentation for authentication and user creation

paths:
  /api/create-user:
    post:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                name:
                  type: string
      responses:
        '200':
          description: User created successfully
        '400':
          description: Missing fields
        '500':
          description: Server error

  /api/authentication:
    post:
      summary: Login an existing user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
        '401':
          description: Unauthorized or Email not confirmed
        '500':
          description: Server error

  /api/set-token:
    post:
      summary: Set a JWT token in cookies after OTP verification
      description: This endpoint is called after OTP is successfully verified. It signs a JWT token and sets it in an HTTP-only cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  description: The email address of the verified user
      responses:
        '200':
          description: Token created and stored in cookie
        '400':
          description: Missing email field
        '500':
          description: Token creation failed